refactor(store): move csf related logic to its own folder
