+++ lib/store/src/csf/getValuesFromArgTypes.ts
@@ -1,0 +1,9 @@
+import { ArgTypes } from '@storybook/csf';
+
+export const getValuesFromArgTypes = (argTypes: ArgTypes = {}) =>
+  Object.entries(argTypes).reduce((acc, [arg, { defaultValue }]) => {
+    if (typeof defaultValue !== 'undefined') {
+      acc[arg] = defaultValue;
+    }
+    return acc;
+  }, {} as ArgTypes);
