+++ lib/store/src/csf/normalizeProjectAnnotations.ts
@@ -1,0 +1,25 @@
+import { AnyFramework, ProjectAnnotations } from '@storybook/csf';
+import { inferControls, NormalizedProjectAnnotations, normalizeInputTypes } from '..';
+import { inferArgTypes } from '../inferArgTypes';
+
+export function normalizeProjectAnnotations<TFramework extends AnyFramework>({
+  argTypes,
+  globalTypes,
+  argTypesEnhancers,
+  ...annotations
+}: ProjectAnnotations<TFramework>): NormalizedProjectAnnotations<TFramework> {
+  return {
+    ...(argTypes && { argTypes: normalizeInputTypes(argTypes) }),
+    ...(globalTypes && { globalTypes: normalizeInputTypes(globalTypes) }),
+    argTypesEnhancers: [
+      ...(argTypesEnhancers || []),
+      inferArgTypes,
+      // inferControls technically should only run if the user is using the controls addon,
+      // and so should be added by a preset there. However, as it seems some code relies on controls
+      // annotations (in particular the angular implementation's `cleanArgsDecorator`), for backwards
+      // compatibility reasons, we will leave this in the store until 7.0
+      inferControls,
+    ],
+    ...annotations,
+  };
+}
