+++ lib/store/src/csf/normalizeStory.test.ts
