+++ lib/store/src/index.ts
@@ -1,12 +1,12 @@
 export { StoryStore } from './StoryStore';
 export { combineParameters } from './parameters';
 export { filterArgTypes } from './filterArgTypes';
-export { normalizeInputTypes } from './normalizeInputTypes';
 export type { PropDescriptor } from './filterArgTypes';
 export { inferControls } from './inferControls';
 
 export * from './types';
 
+export * from './csf';
 export * from './hooks';
 export * from './decorators';
 export * from './args';
